# api/openapi.yaml
openapi: 3.0.0
info:
  title: Content Sharing Platform API
  description: |
    A comprehensive REST API for a content sharing platform that allows users to share, discover, and rate media content.
    
    ## Features
    - JWT-based authentication
    - User registration and login
    - Media content management (CRUD)
    - Rating system (1-5 stars)
    - Advanced filtering and search
    - Pagination support
    - Rating statistics
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. After logging in or registering, include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
...
paths:
...
  /api/content:
    get:
      tags:
        - Content
      summary: List all content
      description: Get a paginated list of media content with optional filtering and sorting
      operationId: listContent
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: Items per page
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: category
          in: query
          description: Filter by category
          schema:
            $ref: '#/components/schemas/MediaCategory'
        - name: user_id
          in: query
          description: Filter by creator user ID
          schema:
            type: integer
        - name: search
          in: query
          description: Search in title and description
          schema:
            type: string
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            default: created_at
            enum: [created_at, title, category, user_id]
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            default: desc
            enum: [asc, desc]
      responses:
        '200':
          description: List of content
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaContent'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Content
      summary: Create content
      description: Create a new media content item (requires authentication)
      operationId: createContent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaContentCreate'
      responses:
        '201':
          description: Content created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Content created successfully
                  content:
                    $ref: '#/components/schemas/MediaContent'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/content/{media_id}:
    get:
      tags:
        - Content
      summary: Get content by ID
      description: Get detailed information about a specific media content
      operationId: getContent
      parameters:
        - name: media_id
          in: path
          required: true
          description: Media content ID
          schema:
            type: integer
      responses:
        '200':
          description: Content details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaContentDetail'
        '404':
          description: Content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Content
      summary: Update content
      description: Update an existing media content (only by creator)
      operationId: updateContent
      security:
        - BearerAuth: []
      parameters:
        - name: media_id
          in: path
          required: true
          description: Media content ID
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaContentUpdate'
      responses:
        '200':
          description: Content updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Content updated successfully
                  content:
                    $ref: '#/components/schemas/MediaContent'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the content creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Content
      summary: Delete content
      description: Delete a media content (only by creator)
      operationId: deleteContent
      security:
        - BearerAuth: []
      parameters:
        - name: media_id
          in: path
          required: true
          description: Media content ID
          schema:
            type: integer
      responses:
        '200':
          description: Content deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Content deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the content creator
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Content not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
    MediaCategory:
      type: string
      enum:
        - video
        - image
        - audio
        - document
        - other
      description: Category of media content
...
    User:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        username:
          type: string
          example: johndoe
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
...
    MediaContentCreate:
      type: object
      required:
        - title
        - category
        - content_url
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Amazing Sunset Video
          description: Content title (max 200 characters)
        description:
          type: string
          example: A beautiful sunset captured at the beach
          description: Optional content description
        category:
          $ref: '#/components/schemas/MediaCategory'
        thumbnail_url:
          type: string
          format: uri
          example: https://example.com/thumbnails/sunset.jpg
          description: Optional thumbnail URL
        content_url:
          type: string
          format: uri
          example: https://example.com/videos/sunset.mp4
          description: URL to the actual content
    MediaContentUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Updated Sunset Video
        description:
          type: string
          example: Updated description
        category:
          $ref: '#/components/schemas/MediaCategory'
        thumbnail_url:
          type: string
          format: uri
          example: https://example.com/thumbnails/sunset-new.jpg
        content_url:
          type: string
          format: uri
          example: https://example.com/videos/sunset-hd.mp4
    MediaContent:
      type: object
      properties:
        media_id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        title:
          type: string
          example: Amazing Sunset Video
        description:
          type: string
          example: A beautiful sunset captured at the beach
        category:
          $ref: '#/components/schemas/MediaCategory'
        thumbnail_url:
          type: string
          format: uri
          example: https://example.com/thumbnails/sunset.jpg
        content_url:
          type: string
          format: uri
          example: https://example.com/videos/sunset.mp4
        created_at:
          type: string
          format: date-time
          example: "2024-01-15T14:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-01-15T14:30:00Z"
    MediaContentDetail:
      allOf:
        - $ref: '#/components/schemas/MediaContent'
        - type: object
          properties:
            creator:
              $ref: '#/components/schemas/User'
            rating_stats:
              type: object
              properties:
                total_ratings:
                  type: integer
                  example: 42
                average_rating:
                  type: number
                  format: float
                  example: 4.5
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
          description: Current page number
        per_page:
          type: integer
          example: 10
          description: Items per page
        total_items:
          type: integer
          example: 42
          description: Total number of items
        total_pages:
          type: integer
          example: 5
          description: Total number of pages
    Error:
      type: object
      properties:
        error:
          type: string
          example: Validation error
          description: Error message
        message:
          type: string
          example: Invalid input data
          description: Detailed error description